1.1

+++ src/main/java/com/jpr/app/domain/FactScrapIssued.java
@@ -30,7 +30,7 @@ public class FactScrapIssued {
 	@OnDelete(action =OnDeleteAction.CASCADE)
 	private DimDate dimDate;
 	
-	private int quantity;
+	private Double quantity;
 
 	public ScrapRecvPk getId() {
 		return id;
@@ -56,13 +56,15 @@ public class FactScrapIssued {
 		this.dimDate = dimDate;
 	}
 
-	public int getQuantity() {
+	public Double getQuantity() {
 		return quantity;
 	}
 
-	public void setQuantity(int quantity) {
+	public void setQuantity(Double quantity) {
 		this.quantity = quantity;
 	}
+
+
